# docs/swagger.yaml
basePath: /
definitions:
  badresponse.ResponseError:
    properties:
      reason:
        type: string
      time:
        type: string
    type: object
...
  model.Blog:
    properties:
      authorId:
        example: xxxxxxxxxxxxx==
        type: string
      categories:
        items:
          type: string
        type: array
      content:
        example: |-
          xxxx
          xxxx
        type: string
      cover:
        example: https://xxx/xxx
        type: string
      description:
        example: mouse ❤ monkey
        type: string
      entityInfo:
        $ref: '#/definitions/model.Entity'
      id:
        example: xxxxxxxxxxxxx==
        type: string
      keyWords:
        example:
        - xxx
        - xxx
        items:
          type: string
        type: array
      title:
        example: mouse ❤ monkey
        type: string
    type: object
...
  model.Entity:
    properties:
      createTime:
        example: "2020-10-1"
        type: string
      is_deleted:
        type: boolean
      updateTime:
        example: "2020-10-1"
        type: string
    type: object
...
host: localhost:5001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Mo2 server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mo2
  version: "1.0"
paths:
...
  /api/blogs/{id}:
    delete:
      consumes:
      - application/json
      description: delete by id path(draft/blog)
      parameters:
      - description: bool true
        in: query
        name: draft
        required: true
        type: boolean
      - description: string xxxxxxxx
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "204":
          description: ""
        "304":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: 彻底删除blog
      tags:
      - blogs
  /api/blogs/{operation}/{id}:
    put:
      consumes:
      - application/json
      description: restore by path
      parameters:
      - description: bool true
        in: query
        name: draft
        required: true
        type: boolean
      - description: recycle/restore
        in: path
        name: operation
        type: string
      - description: Blog id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: restore Blog
      tags:
      - blogs
...
  /api/blogs/find/id:
    get:
      consumes:
      - application/json
      parameters:
      - description: bool true
        in: query
        name: draft
        type: boolean
      - description: string xxxxxxxx
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Blog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find Blog by id
      tags:
      - blogs
  /api/blogs/find/own:
    get:
      consumes:
      - application/json
      parameters:
      - description: bool true
        in: query
        name: draft
        type: boolean
      - description: bool default false
        in: query
        name: deleted
        type: boolean
      - description: int 0
        in: query
        name: page
        type: integer
      - description: int 5
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "204":
          description: No Content
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find Blog
      tags:
      - blogs
  /api/blogs/find/userId:
    get:
      consumes:
      - application/json
      parameters:
      - description: bool true
        in: query
        name: draft
        type: boolean
      - description: bool default false
        in: query
        name: deleted
        type: boolean
      - description: string xxxxxxxx
        in: query
        name: id
        type: string
      - description: int 0
        in: query
        name: page
        type: integer
      - description: int 5
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find Blog by userid
      tags:
      - blogs
  /api/blogs/publish:
    post:
      consumes:
      - application/json
      description: add by json
      parameters:
      - description: bool true
        in: query
        name: draft
        type: boolean
      - description: Add blog
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.Blog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Blog'
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: Publish Blog
      tags:
      - blogs
  /api/blogs/query:
    get:
      description: find
      parameters:
      - description: bool default false
        in: query
        name: draft
        type: boolean
      - description: bool default false
        in: query
        name: deleted
        type: boolean
      - description: int 0
        in: query
        name: page
        type: integer
      - description: int 5
        in: query
        name: pageSize
        type: integer
      - description: aaaa
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "204":
          description: No Content
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find all Blogs
      tags:
      - blogs
...
  /api/relation/blogs/{type}/{ID}:
    get:
      consumes:
      - application/json
      description: 根据type返回不同结果：[category] 所有category包含的blog
      parameters:
      - description: find by category
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find blogs by given type
      tags:
      - relation
...
swagger: "2.0"